+++ sourcetracker/_util.py
@@ -12,12 +12,12 @@ import pandas as pd
 
 
 def parse_sample_metadata(f):
-    """ Parse QIIME 1-like sample metadata file
+    """Parse QIIME 1-like sample metadata file.
 
     Parameters
     ----------
     f : file handle
-        The sample metadata to be parse.
+        The sample metadata to be parsed.
 
     Returns
     -------
@@ -30,3 +30,21 @@ def parse_sample_metadata(f):
     sample_metadata.set_index(sample_metadata.columns[0], drop=True,
                               append=False, inplace=True)
     return sample_metadata
+
+
+def biom_to_df(biom_table):
+    '''Turn biom table into dataframe.
+
+    Parameters
+    ----------
+    biom_table : biom.table.Table
+        Biom table.
+
+    Returns
+    -------
+    feature_table : pd.DataFrame
+        Contingency table with rows, columns = samples, features.
+    '''
+    return pd.DataFrame(biom_table._data.toarray().T,
+                        index=biom_table.ids(axis='sample'),
+                        columns=biom_table.ids(axis='observation'))